+++ src/domain/auth/components/login/Login.tsx
@@ -1,29 +1,22 @@
 import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { differenceInYears } from 'date-fns';
-import { connect, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { useMatomo } from '@datapunt/matomo-tracker-react';
 import { Button, IconLinkExternal } from 'hds-react';
 
 import LinkButton from '../../../../common/components/linkButton/LinkButton';
 import Text from '../../../../common/components/text/Text';
-import { AuthState, resetError, isAuthenticatedSelector } from '../../redux';
-import { RootState } from '../../../../redux/rootReducer';
+import { isAuthenticatedSelector } from '../../redux';
 import authenticate from '../../authenticate';
 import PageContentWithHostingBox from '../../../../common/components/layout/PageContentWithHostingBox';
 import styles from './Login.module.css';
 import BirthdateForm from '../birthdateForm/BirthdateForm';
-import NotificationComponent from '../../../../common/components/notification/NotificationComponent';
 import authConstants from '../../constants/authConstants';
 import ageConstants from '../../../youthProfile/constants/ageConstants';
 import PageWrapper from '../../../../common/components/wrapper/PageWrapper';
 
-type Props = {
-  resetError: () => void;
-  auth: AuthState;
-};
-
-function Login(props: Props) {
+function Login() {
   const [showManualRegistration, setShowManualRegistration] = useState(false);
   const { t } = useTranslation();
   const { trackEvent } = useMatomo();
@@ -112,20 +105,9 @@ function Login(props: Props) {
             </div>
           )}
         </div>
-
-        <NotificationComponent
-          show={Boolean(props.auth.error)}
-          onClose={() => props.resetError()}
-        />
       </PageContentWithHostingBox>
     </PageWrapper>
   );
 }
 
-const mapStateToProps = (state: RootState) => {
-  return {
-    auth: state.auth,
-  };
-};
-
-export default connect(mapStateToProps, { resetError })(Login);
+export default Login;